Extract bundle info code
